+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+vx.y.z - TBD
+------------
+
+* [#270](https://github.com/godaddy/tartufo/issues/270) - When no refs/branches
+  are found locally, tartufo will now scan the repo HEAD as a single commit,
+  effectively scanning the entire codebase at once.
+
 v3.0.0-alpha.1 - 11 November 2021
 ---------------------------------